Move D1 cooldown calculation into D1CharacterStats
